Add a loadgen tool.
