+++ tools/loadgen/.dockerignore
@@ -1,0 +1,5 @@
+target
+*.iml
+.idea
+.git
+
